% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday.R
\name{birthday}
\alias{birthday}
\title{Title: birthday}
\usage{
birthday(x)
}
\arguments{
\item{x}{The number of people in the group}
}
\value{
A vector of probabilities corresponding to the input
}
\description{
Title: birthday
}
\examples{
# Calculate probabilities for default class sizes 20-25 birthday()
}
