% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title: Binomial Simulation}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
Table of proportions of successes
}
\description{
Title: Binomial Simulation
}
\examples{
\dontrun{mybin(iter,n,p)}
}
