% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title Bootstrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations}

\item{x}{Sample data}

\item{fun}{Function to calculate statistic}

\item{alpha}{Alpha level for confidence interval calculation}

\item{cx}{Text scaling}

\item{...}{Additional graphical parameters}
}
\value{
List containing confidence interval, function used, and original data
}
\description{
Title Bootstrap Confidence Interval Function
}
\examples{
\dontrun{
  myboot2(x=rnorm(20))
}
}
