% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Title Sampling Distribution of the Sample Mean (Uniform)}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Integer. The sample size for each iteration.}

\item{iter}{Integer. The number of iterations (samples) to simulate}

\item{a}{Numeric. Lower bound of the Uniform distribution}

\item{b}{Numeric. Upper bound of the Uniform distribution}
}
\value{
a numeric vector of length iter containing the simulated sample means
}
\description{
Title Sampling Distribution of the Sample Mean (Uniform)
}
\examples{
\dontrun{
  w <- mycltu(n = 20, iter = 10000, a = 0, b = 10)
  mean(w)
  var(w)
}
}
