% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function that computes log-likelihood for each data point}

\item{x}{Numeric vector of observed data}

\item{param}{Numeric vector of parameter values to test}

\item{...}{Additional graphical parameters}
}
\value{
A list containing:
\item{i}{Index of maximum likelihood estimate}
\item{parami}{Parameter value at maximum likelihood}
\item{yi}{Log-likelihood value at maximum}
\item{slope}{Slope around the maximum (if calculable)}
}
\description{
Maximum Likelihood Estimation
}
\examples{
# Binomial example
data <- c(9,9,1,9,9,9)
result <- mymaxlik(x=data, param=seq(0,1,length=1000),
                   lfun=function(x,p) log(dbinom(x,prob=p,size=10)))
}
