% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatterhist}
\alias{scatterhist}
\title{A scatter plot with marginal histograms}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{a quantitative vector}

\item{y}{a quantitative vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A scatter plot of x vs y with marginal histograms
}
\description{
A scatter plot with marginal histograms
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH, WEIGHT, xlab = "LENGTH"))}
}
